provide full team keyring when deserializing and when admitting members
The previous strategy was to always decrypt a link at a time, reducing as we go, so that as the team keys are rotated we can get them from lockboxes. That's still what we want to do when communicating with an existing member. But it doesn't always work (a) when rehydrating a team from storage or (b) when joining a team: We'll never have lockboxes for key rotations that happened *before we joined.* So it's better in those situations to always use a keyring containing the full history of team keys.