Team: make sure joined device has all team keys
Without this, I was finding that a joined device only had access to the most recent generation of team keys. When the device would try to instantiate the team from storage, it couldn't decrypt the root link if that link was encrypted with an older generation. It's possible the joined device should hang out and wait for the `admit member` link or some other way of getting all the team keys, but this worked and I've been banging my head for several hours