Improve performance of memoized validator functions.
The cache key was based on a hash of the entire graph. This hash was being
generated each time a link was validated. On my system, validating each link
took around 100ms with a graph of around 2000 links. With this change it appears
that validating each link takes less than 1ms.